// types/cartItem.ts
export interface CartItem {
    specie_name: string;
    breed_name: string;
    category_id: string;
    category_name: string;
    category_description: string;
    color_id: string;
    color_label: string;
    product_id: string;
    product_name: string;
    description: string;
    stokc_id: number;
    stock: number;
    stock_updated: string;
    price: number;
    discount: number;
    image_url: string;
    created_at: string;
    updated_at: string;
    is_saved: any;
    quantity: number;
}
  